svn merge -r3043:3064 ^/branches/versions
git-svn-id: svn+ssh://rubyforge.org/var/svn/myexperiment/trunk@3065 4f7b4649-9ca9-420a-81b3-55a188e4d46f

+++ lib/rest.rb
@@ -684,7 +684,7 @@ def rest_resource_uri(ob)
     when 'Attribution';     return nil
     when 'Tagging';         return nil
 
-    when 'Workflow::Version'; return "#{rest_resource_uri(ob.workflow)}?version=#{ob.version}"
+    when 'WorkflowVersion'; return "#{rest_resource_uri(ob.workflow)}?version=#{ob.version}"
   end
 
   raise "Class not processed in rest_resource_uri: #{ob.class.to_s}"
@@ -728,7 +728,7 @@ def rest_access_uri(ob)
     when 'Creditation';     return "#{base}/credit.xml?id=#{ob.id}"
     when 'Attribution';     return nil
 
-    when 'Workflow::Version'; return "#{base}/workflow.xml?id=#{ob.workflow.id}&version=#{ob.version}"
+    when 'WorkflowVersion'; return "#{base}/workflow.xml?id=#{ob.workflow.id}&version=#{ob.version}"
   end
 
   raise "Class not processed in rest_access_uri: #{ob.class.to_s}"
@@ -752,7 +752,7 @@ def rest_object_tag_text(ob)
     when 'Download';               return 'download'
     when 'PackContributableEntry'; return rest_object_tag_text(ob.contributable)
     when 'PackRemoteEntry';        return 'external'
-    when 'Workflow::Version';      return 'workflow'
+    when 'WorkflowVersion';        return 'workflow'
     when 'Comment';                return 'comment'
     when 'Bookmark';               return 'favourite'
     when 'ContentType';            return 'type'
@@ -784,7 +784,7 @@ def rest_object_label_text(ob)
     when 'Download';               return ''
     when 'PackContributableEntry'; return rest_object_label_text(ob.contributable)
     when 'PackRemoteEntry';        return ob.title     
-    when 'Workflow::Version';      return ob.title
+    when 'WorkflowVersion';        return ob.title
     when 'ContentType';            return ob.title
     when 'License';                return ob.title
     when 'CurationEvent';          return ob.category
@@ -1090,11 +1090,14 @@ def workflow_aux(action, opts = {})
       return rest_response(500, :reason => "Unable to extract metadata")
     end
 
-    success = if (action == 'create' and opts[:query]['id'])
-      ob.save_as_new_version(revision_comment)
-    else
-      ob.save
+    new_version = action == 'create' && opts[:query]['id']
+
+    if new_version
+      ob.preview = nil
+      ob[:revision_comments] = revision_comment
     end
+
+    success = ob.save
 
     return rest_response(400, :object => ob) unless success
 
@@ -1137,7 +1140,11 @@ def file_aux(action, opts = {})
   case action
     when 'create':
       return rest_response(401, :reason => "Not authorised to create a file") unless Authorization.is_authorized_for_type?('create', 'Blob', opts[:user], nil)
-      ob = Blob.new(:contributor => opts[:user])
+      if opts[:query]['id']
+        ob, error = obtain_rest_resource('Blob', opts[:query]['id'], opts[:query]['version'], opts[:user], action)
+      else
+        ob = Blob.new(:contributor => opts[:user])
+      end
     when 'read', 'update', 'destroy':
       ob, error = obtain_rest_resource('Blob', opts[:query]['id'], opts[:query]['version'], opts[:user], action)
     else
@@ -1148,18 +1155,21 @@ def file_aux(action, opts = {})
 
   if action == "destroy"
 
+    return rest_response(400, :reason => "Cannot delete individual versions") if opts[:query]['version']
+      
     ob.destroy
 
   else
 
     data = LibXML::XML::Parser.string(request.raw_post).parse
 
-    title        = parse_element(data, :text,   '/file/title')
-    description  = parse_element(data, :text,   '/file/description')
-    license_type = parse_element(data, :text,   '/file/license-type')
-    type         = parse_element(data, :text,   '/file/type')
-    content_type = parse_element(data, :text,   '/file/content-type')
-    content      = parse_element(data, :binary, '/file/content')
+    title            = parse_element(data, :text,   '/file/title')
+    description      = parse_element(data, :text,   '/file/description')
+    license_type     = parse_element(data, :text,   '/file/license-type')
+    type             = parse_element(data, :text,   '/file/type')
+    content_type     = parse_element(data, :text,   '/file/content-type')
+    content          = parse_element(data, :binary, '/file/content')
+    revision_comment = parse_element(data, :text,   '/workflow/revision-comment')
 
     permissions  = data.find_first('/file/permissions')
 
@@ -1211,17 +1221,28 @@ def file_aux(action, opts = {})
 
     ob.content_blob = ContentBlob.new(:data => content) if content
 
-    if not ob.save
-      return rest_response(400, :object => ob)
+    new_version = action == 'create' && opts[:query][:id]
+
+    if new_version
+      ob[:revision_comments] = revision_comment
     end
 
-    if ob.contribution.policy.nil?
-      ob.contribution.policy = create_default_policy(opts[:user])
-      ob.contribution.save
-    end
+    success = ob.save
 
-    update_permissions(ob, permissions)
+    return rest_response(400, :object => ob) unless success
+
+    if opts[:query]['version'].nil?
+
+      if ob.contribution.policy.nil?
+        ob.contribution.policy = create_default_policy(opts[:user])
+        ob.contribution.save
+      end
+
+      update_permissions(ob, permissions)
+    end
   end
+
+  ob = ob.versioned_resource if ob.respond_to?("versioned_resource")
 
   rest_get_request(ob, opts[:user], { "id" => ob.id.to_s })
 end
@@ -2236,10 +2257,10 @@ end
 # Call dispatcher
 
 def rest_call_request(opts)
-  begin
+# begin
     send(opts[:rules]['Function'], opts)
-  rescue
-    return rest_response(500)
-  end
+# rescue
+#   return rest_response(500)
+# end
 end
